feat: [BE] Teacher CRUD Example

See merge request s10-webmobile1-sub2/S10P12C206!3

+++ backend/src/main/java/com/ssafy/domain/classroom/repository/GroupRepository.java
@@ -2,5 +2,5 @@ package com.ssafy.domain.classroom.repository;
 
 import com.ssafy.domain.classroom.entity.Group;
 import org.springframework.data.jpa.repository.JpaRepository;
-public interface GroupRepository extends JpaRepository<Group, Integer> {
+public interface GroupRepository extends JpaRepository<Group, Long> {
 }
